Organize the code into different files
