+++ src-tauri/src/xap/hid.rs
@@ -1,0 +1,5 @@
+mod client;
+mod device;
+
+pub use client::*;
+pub use device::*;
