+++ src-tauri/src/xap/protocol/routes.rs
@@ -1,0 +1,21 @@
+mod route_imports;
+mod xap;
+mod qmk;
+mod rgblight;
+
+pub use xap::{
+    XAPVersion, XAPVersionQuery,
+    XAPCapabilities, XAPCapabilitiesQuery,
+    XAPEnabledSubsystems, XAPEnabledSubsystemsQuery,
+    XAPSecureStatus, XAPSecureStatusQuery,
+    XAPSecureStatusUnlock,
+    XAPSecureStatusLock,
+};
+pub use qmk::{
+    QMKVersion, QMKVersionQuery,
+    QMKCapabilities, QMKCapabilitiesQuery,
+};
+pub use rgblight::{
+    RGBLightConfig, RGBLightConfigQuery,
+    RGBLightConfigCommand,
+};
